excel_project/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.1
colorama==0.4.4
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.16.0
dnspython==2.2.1
et_xmlfile==2.0.0
exrex==0.11.0
fire==0.4.0
future==0.18.2
idna==3.3
IMAPClient==3.0.1
loguru==0.6.0
mysqlclient==2.2.7
numpy==2.2.1
openpyxl==3.1.5
pandas==2.2.3
PyMySQL==1.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.28.1
setuptools==75.8.0
six==1.16.0
soupsieve==2.3.2
SQLAlchemy==1.3.22
sqlparse==0.5.3
tenacity==8.0.1
termcolor==1.1.0
tqdm==4.64.0
treelib==1.6.1
tzdata==2024.2
urllib3==1.26.9
wheel==0.45.1
win32-setctime==1.1.0
